+++ src/components/sections/elements/Lineage.vue
@@ -1,6 +1,8 @@
 <script setup lang="ts">
 import { ref, watch } from 'vue'
 
+import { Stability, SectionType } from '@/types/enum'
+
 import SectionBorder from '@/components/bases/SectionBorder.vue'
 import SectionTitle from '@/components/bases/SectionTitle.vue'
 import Markdown from '@/components/bases/Markdown.vue'
@@ -20,8 +22,14 @@ watch(
 </script>
 
 <template>
-  <SectionBorder>
-    <SectionTitle version="1.2" anchor="lineage" title="Lineage" />
+  <SectionBorder :type="SectionType.Element">
+    <SectionTitle
+      :type="SectionType.Element"
+      version="1.3"
+      :stability="Stability.Stable"
+      anchor="lineage"
+      title="Lineage"
+    />
     <Markdown input-id="lineage-input" @update:input="(event: string) => (statement = event)" />
   </SectionBorder>
 </template>