+++ src/components/sections/elements/__tests__/Downloads.spec.ts
@@ -22,9 +22,9 @@ describe('Downloads', () => {
   let tocItemsDiv: HTMLDivElement
 
   beforeEach(() => {
-    // TOC link in section title will be teleported into a '#toc-items' element so create a fake one to stop warnings
+    // TOC link in section title will be teleported into a '#toc-items-element' element so create a fake one to stop warnings
     tocItemsDiv = document.createElement('div')
-    tocItemsDiv.id = 'toc-items'
+    tocItemsDiv.id = 'toc-items-element'
     document.body.appendChild(tocItemsDiv)
   })
 
@@ -106,9 +106,9 @@ describe('Downloads [Integration]', () => {
   let tocItemsDiv: HTMLDivElement
 
   beforeEach(() => {
-    // TOC link in section title will be teleported into a '#toc-items' element so create a fake one to stop warnings
+    // TOC link in section title will be teleported into a '#toc-items-element' element so create a fake one to stop warnings
     tocItemsDiv = document.createElement('div')
-    tocItemsDiv.id = 'toc-items'
+    tocItemsDiv.id = 'toc-items-element'
     document.body.appendChild(tocItemsDiv)
   })