+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Changed
 
+* Refactoring Table of Contents and SectionTitle/SectionBorder base components into types
 * Updating to ISO v4 and MAGIC Discovery v1 record schemas
 * Reorganise static components separately from base components
 * upgrading dependencies